move utility function outside of script
